feat: renamed main file to 'main.py'; moved database functions to a seperate file 'database_functions.py'; moved menu system to a seperate file 'menu_screens.py'; completely removed application logging, rework is necessacy; removed DatabaseSession class, as it was unnecessary and was causing complications with the new file system
